high angle support
Its shoots like 2 blocks short if target is more then ~200 blocks away and shooting high angle idk why.

+++ src/main/java/com/happysg/radar/compat/cbc/CannonUtil.java
@@ -42,10 +42,10 @@ public class CannonUtil {
 
             Block block = containedBlockInfo.state().getBlock();
             if (block instanceof BigCannonPropellantBlock propellantBlock) {
-                chargePower += propellantBlock.getChargePower(containedBlockInfo);
+                chargePower += (int) propellantBlock.getChargePower(containedBlockInfo);
             } else if (block instanceof ProjectileBlock<?> projectileBlock) {
                 AbstractBigCannonProjectile projectile = projectileBlock.getProjectile(level, Collections.singletonList(containedBlockInfo));
-                chargePower += projectile.addedChargePower();
+                chargePower += (int) projectile.addedChargePower();
             }
         }
         return chargePower;
